backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.5.0
pydantic==2.8.0
python-dotenv==1.0.1
pymongo==4.8.0